requirements.txt
Flask
cryptography
pycryptodome
numpy
